refactor: merge process_actions and main in __main__.py
