+++ incipyt/__init__.py
@@ -1,2 +1,2 @@
 from incipyt import actions, hooks
-from incipyt.system import Environment, Hierarchy, process_actions
+from incipyt.os import Environment, Hierarchy
